# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LubisiaTech Solutions API
servers:
  - url: /api
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    patch:
      operationId: updateService
      tags: [services]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
      responses:
        "204":
          description: Deleted

  # ── Products ──────────────────────────────────────────────────────────────
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      parameters:
        - {name: category, in: query, required: false, schema: {type: string}}
        - {name: inStock, in: query, required: false, schema: {type: boolean}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
      responses:
        "204":
          description: Deleted

  # ── Orders ────────────────────────────────────────────────────────────────
...
  /settings/{key}:
    get:
      operationId: getSetting
      tags: [settings]
      parameters:
        - {name: key, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSetting"
        "404":
          description: Not found
    put:
      operationId: upsertSetting
      tags: [settings]
      parameters:
        - {name: key, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSetting"

  # ── Affiliate Programs ────────────────────────────────────────────────────
...
  /website-templates:
    get:
      operationId: listWebsiteTemplates
      tags: [websiteTemplates]
      parameters:
        - {name: category, in: query, required: false, schema: {type: string}}
        - {name: featured, in: query, required: false, schema: {type: boolean}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebsiteTemplate"
    post:
      operationId: createWebsiteTemplate
      tags: [websiteTemplates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteTemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebsiteTemplate"
  /website-templates/{id}:
    get:
      operationId: getWebsiteTemplate
      tags: [websiteTemplates]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebsiteTemplate"
        "404":
          description: Not found
    patch:
      operationId: updateWebsiteTemplate
      tags: [websiteTemplates]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteTemplateUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebsiteTemplate"
    delete:
      operationId: deleteWebsiteTemplate
      tags: [websiteTemplates]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Service:
      type: object
      required: [id, name, description, category, createdAt]
      properties:
        id: {type: integer}
        name: {type: string}
        description: {type: string}
        category: {type: string}
        price: {type: ["number", "null"]}
        imageUrl: {type: ["string", "null"]}
        featured: {type: boolean}
        createdAt: {type: string}
    ServiceInput:
      type: object
      required: [name, description, category]
      properties:
        name: {type: string}
        description: {type: string}
        category: {type: string}
        price: {type: ["number", "null"]}
        imageUrl: {type: ["string", "null"]}
        featured: {type: boolean}
    ServiceUpdate:
      type: object
      properties:
        name: {type: string}
        description: {type: string}
        category: {type: string}
        price: {type: ["number", "null"]}
        imageUrl: {type: ["string", "null"]}
        featured: {type: boolean}
    Product:
      type: object
      required: [id, name, description, category, price, stock, createdAt]
      properties:
        id: {type: integer}
        name: {type: string}
        description: {type: string}
        category: {type: string}
        price: {type: number}
        stock: {type: integer}
        imageUrl: {type: ["string", "null"]}
        featured: {type: boolean}
        createdAt: {type: string}
    ProductInput:
      type: object
      required: [name, description, category, price, stock]
      properties:
        name: {type: string}
        description: {type: string}
        category: {type: string}
        price: {type: number}
        stock: {type: integer}
        imageUrl: {type: ["string", "null"]}
        featured: {type: boolean}
    ProductUpdate:
      type: object
      properties:
        name: {type: string}
        description: {type: string}
        category: {type: string}
        price: {type: number}
        stock: {type: integer}
        imageUrl: {type: ["string", "null"]}
        featured: {type: boolean}
...
    SiteSetting:
      type: object
      required: [id, key, value, label, group, updatedAt]
      properties:
        id: {type: integer}
        key: {type: string}
        value: {type: string}
        label: {type: string}
        group: {type: string}
        updatedAt: {type: string}
    SiteSettingInput:
      type: object
      required: [value, label, group]
      properties:
        value: {type: string}
        label: {type: string}
        group: {type: string}
...
    WebsiteTemplate:
      type: object
      required: [id, name, description, category, price, active, featured, createdAt]
      properties:
        id: {type: integer}
        name: {type: string}
        description: {type: string}
        category: {type: string}
        price: {type: number}
        previewUrl: {type: ["string", "null"]}
        screenshotUrl: {type: ["string", "null"]}
        features: {type: string}
        techStack: {type: ["string", "null"]}
        deliveryDays: {type: integer}
        active: {type: boolean}
        featured: {type: boolean}
        createdAt: {type: string}
    WebsiteTemplateInput:
      type: object
      required: [name, description, category, price]
      properties:
        name: {type: string}
        description: {type: string}
        category: {type: string}
        price: {type: number}
        previewUrl: {type: ["string", "null"]}
        screenshotUrl: {type: ["string", "null"]}
        features: {type: string}
        techStack: {type: ["string", "null"]}
        deliveryDays: {type: integer}
        active: {type: boolean}
        featured: {type: boolean}
    WebsiteTemplateUpdate:
      type: object
      properties:
        name: {type: string}
        description: {type: string}
        category: {type: string}
        price: {type: number}
        previewUrl: {type: ["string", "null"]}
        screenshotUrl: {type: ["string", "null"]}
        features: {type: string}
        techStack: {type: ["string", "null"]}
        deliveryDays: {type: integer}
        active: {type: boolean}
        featured: {type: boolean}