# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LubisiaTech Solutions API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description: string;
  category: string;
  price: number;
  stock: number;
  /** @nullable */
  imageUrl?: string | null;
  featured?: boolean;
}