# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LubisiaTech Solutions API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  description: string;
  category: string;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  featured?: boolean;
  createdAt: string;
}