# lib/api-zod/src/generated/types/siteSettingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LubisiaTech Solutions API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettingInput {
  value: string;
  label: string;
  group: string;
}